api/requirements.txt
google-generativeai
Pillow
razorpay
firebase-admin
